backend/requirements.txt
dnspython==2.4.2 